feat(shim-kvm): move consts to lib.rs
